Fixed typos in some of the tile wrappers

+++ geopyspark-backend/geotrellis/src/main/scala/geopyspark/geotrellis/DoubleArrayTileWrapper.scala
@@ -11,9 +11,9 @@ object DoubleArrayTileWrapper extends Wrapper[DoubleArrayTile] {
 
   def testRdd(sc: SparkContext): RDD[DoubleArrayTile] = {
     val arr = Array(
-      DoubleArrayTile(Array[Byte](0, 0, 1, 1), 2, 2),
-      DoubleArrayTile(Array[Byte](1, 2, 3, 4), 2, 2),
-      DoubleArrayTile(Array[Byte](5, 6, 7, 8), 2, 2))
+      DoubleArrayTile(Array[Double](0, 0, 1, 1), 2, 2),
+      DoubleArrayTile(Array[Double](1, 2, 3, 4), 2, 2),
+      DoubleArrayTile(Array[Double](5, 6, 7, 8), 2, 2))
     println("\n\n\n")
     println("THESE ARE THE ORIGINAL DOUBLEARRAYTILES")
     arr.foreach(println)
