+++ geopyspark-backend/geotrellis/src/main/scala/geopyspark/geotrellis/FloatArrayTileWrapper.scala
@@ -11,9 +11,9 @@ object FloatArrayTileWrapper extends Wrapper[FloatArrayTile] {
 
   def testRdd(sc: SparkContext): RDD[FloatArrayTile] = {
     val arr = Array(
-      FloatArrayTile(Array[Byte](0, 0, 1, 1), 2, 2),
-      FloatArrayTile(Array[Byte](1, 2, 3, 4), 2, 2),
-      FloatArrayTile(Array[Byte](5, 6, 7, 8), 2, 2))
+      FloatArrayTile(Array[Float](0, 0, 1, 1), 2, 2),
+      FloatArrayTile(Array[Float](1, 2, 3, 4), 2, 2),
+      FloatArrayTile(Array[Float](5, 6, 7, 8), 2, 2))
     println("\n\n\n")
     println("THESE ARE THE ORIGINAL FLOATARRAYTILES")
     arr.foreach(println)
